+++ src/main/java/jamdoggie/betterbattletowers/BattleTowerConfig.java
@@ -1,7 +1,7 @@
 package jamdoggie.betterbattletowers;
 
 import jamdoggie.betterbattletowers.mixins.accessor.TomlAccessor;
-import jamdoggie.betterbattletowers.worldgen.LootTable;
+import jamdoggie.betterbattletowers.worldgen.util.LootTable;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.core.block.Block;
 import net.minecraft.core.block.Blocks;
@@ -22,7 +22,7 @@ import java.util.*;
 import static jamdoggie.betterbattletowers.BattleTowerConfigOld.*;
 import static jamdoggie.betterbattletowers.BattleTowerMod.MOD_ID;
 import static jamdoggie.betterbattletowers.BattleTowerMod.LOGGER;
-import static jamdoggie.betterbattletowers.worldgen.LootTable.getDefaultMap;
+import static jamdoggie.betterbattletowers.worldgen.util.LootTable.getDefaultMap;
 import static net.minecraft.core.util.collection.NamespaceID.getPermanent;
 
 public class BattleTowerConfig {
