+++ src/main/java/jamdoggie/betterbattletowers/BattleTowerConfigOld.java
@@ -1,6 +1,6 @@
 package jamdoggie.betterbattletowers;
 
-import jamdoggie.betterbattletowers.worldgen.LootTable;
+import jamdoggie.betterbattletowers.worldgen.util.LootTable;
 import net.minecraft.core.block.Block;
 import net.minecraft.core.block.Blocks;
 import net.minecraft.core.item.IItemConvertible;
@@ -9,7 +9,7 @@ import turniplabs.halplibe.util.ConfigHandler;
 
 import java.util.*;
 
-import static jamdoggie.betterbattletowers.worldgen.LootTable.LootEntry.loot;
+import static jamdoggie.betterbattletowers.worldgen.util.LootTable.LootEntry.loot;
 
 public class BattleTowerConfigOld {
