+++ src/main/java/jamdoggie/betterbattletowers/worldgen/util/TowerProperty.java
@@ -1,4 +1,4 @@
-package jamdoggie.betterbattletowers.worldgen;
+package jamdoggie.betterbattletowers.worldgen.util;
 
 import net.minecraft.core.WeightedRandomBag;
